+++ dev/ui/src/wpVitePreamble.ts
@@ -2,5 +2,5 @@ import RefreshRuntime from "/@react-refresh";
 
 RefreshRuntime.injectIntoGlobalHook(window);
 window.$RefreshReg$ = () => {};
-window.$RefreshSig$ = () => (type) => type;
+window.$RefreshSig$ = () => (type: unknown) => type;
 window.__vite_plugin_react_preamble_installed__ = true;